implements failed sales features (#94)
* refunds users for failed sales
* uses sales state enum to support failed sales
* overridable internal claiming method
* tests sale failed cases
* drops opening time, deduped beneficiary
* drops sale duration restriction for tests
* sets vesting start point to sales closing time
* fixes price computation on SVCSale
* uses sales state enum
* don't settle crowdsale immediately
* adds a helper to claim saleId manually
---------

+++ test/CrowdSaleFuzz.t.sol
@@ -40,7 +40,7 @@ contract CrowdSaleFuzzTest is Test {
             biddingToken: IERC20(address(biddingToken)),
             fundingGoal: fundingGoal,
             salesAmount: salesAmt,
-            closingTime: block.timestamp + 2 hours
+            closingTime: uint64(block.timestamp + 2 hours)
         });
 
         auctionToken.approve(address(crowdSale), salesAmt);
